moved selectedActors to components. corrected path to file.
